Cosmetic changes.
git-svn-id: http://b035a0a.online-server.cloud/svn/IAcf@668 f04c80de-fe56-0410-a3ee-ab8cdec974cc

+++ Include/iocv/COcvBlobProcessorComp.cpp
@@ -91,8 +91,7 @@ bool COcvBlobProcessorComp::CalculateBlobs(
 	findContours(tmpBinaryImage, contours, CV_RETR_LIST, CV_CHAIN_APPROX_NONE);
 
 	// Get found blobs:
-	for (size_t contourIndex = 0; contourIndex < contours.size(); contourIndex++)
-	{
+	for (int contourIndex = 0; contourIndex < int(contours.size()); contourIndex++){
 		cv::Moments moms = cv::moments(cv::Mat(contours[contourIndex]));
 
 		double area = moms.m00;
